requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=2.2.3
numpy>=1.16.4
opencv-python>=4.1.1.26
Pillow>=7.1.2
PyYAML>=5.1.2
requests>=2.22.0
scipy>=1.3.0
tqdm>=4.36.1
gputil==1.4.0
visualdl>=2.2.0

# Logging -------------------------------------
visualdl>=2.2.0
# wandb

# Plotting ------------------------------------
pandas>=1.1.5
seaborn>=0.10.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export

# Extras --------------------------------------
# albumentations>=1.0.3
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# roboflow
thop  # FLOPs computation
